Properly test all usage models
Note: this testing method is not ideal. We start a separate pytest session for that.
That's because pytest-asyncio does something weird which blocks aiohttp client request
The tests do work, but not when mixing with other pytest-asyncio tests, that's why they are separated
If we find a way to fix this this will be changed